+++ src/app/my-info/_components/DeliveryStatus/DeliveryStatus.tsx
@@ -1,19 +1,45 @@
-import { ROUTER } from '@/constants/route';
-import { ChevronIcon } from '@/public/index';
+'use client';
+
+import { useQuery } from '@tanstack/react-query';
 import classNames from 'classnames/bind';
 import Link from 'next/link';
+
+import { getOrdersData } from '@/api/orderAPI';
+import { ROUTER } from '@/constants/route';
+import { ChevronIcon } from '@/public/index';
+import { Order, OrderStatus } from '@/types/OrderTypes';
+
+import { Fragment, useEffect, useState } from 'react';
 import styles from './DeliveryStatus.module.scss';
 
 const cn = classNames.bind(styles);
 
 const DELIVERY_STATUS_LIST = [
-  { label: '결제완료', count: 1 },
-  { label: '배송준비중', count: 0 },
-  { label: '배송중', count: 0 },
-  { label: '배송완료', count: 0 },
+  { label: '결제 완료', status: OrderStatus.PAYMENT_COMPLETED, count: 0 },
+  { label: '배송 준비중', status: OrderStatus.PREPARING, count: 0 },
+  { label: '배송 중', status: OrderStatus.SHIPPING, count: 0 },
+  { label: '배송 완료', status: OrderStatus.DELIVERED, count: 0 },
+  { label: '구매 확정', status: OrderStatus.CONFIRMED, count: 0 },
 ];
 
 export default function DeliveryStatus() {
+  const [deliveryStatusList, setDeliveryStatusList] = useState(DELIVERY_STATUS_LIST);
+
+  const { data: ordersData } = useQuery<{ data: Order[] }>({ queryKey: ['ordersData'], queryFn: getOrdersData });
+
+  const orders = ordersData?.data ?? [];
+
+  useEffect(() => {
+    const updatedStatusList = deliveryStatusList.map((statusItem) => ({
+      ...statusItem,
+      count: orders.filter((order) => order.orderStatus === statusItem.status).length,
+    }));
+    setDeliveryStatusList(updatedStatusList);
+
+    // [deliveryStatusList, orders]로 종속성 넣으면 무한호출 나와요
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [orders]);
+
   return (
     <article className={cn('delivery-status')}>
       <div className={cn('status-header')}>
@@ -23,16 +49,16 @@ export default function DeliveryStatus() {
         </Link>
       </div>
       <ul className={cn('status-list')}>
-        {DELIVERY_STATUS_LIST.map((status, i) => (
-          <>
-            <li key={status.label} className={cn('status-item')}>
+        {deliveryStatusList.map((status, i) => (
+          <Fragment key={status.label}>
+            <li className={cn('status-item')}>
               <span className={cn('status-count', { active: status.count > 0 })}>{status.count}</span>
               <span className={cn('status-label', { active: status.count > 0 })}>{status.label}</span>
             </li>
-            {i < DELIVERY_STATUS_LIST.length - 1 && (
+            {i < deliveryStatusList.length - 1 && (
               <ChevronIcon className={cn('status-icon', { 'active-icon': status.count > 0 })} />
             )}
-          </>
+          </Fragment>
         ))}
       </ul>
     </article>
